chore: Display information extension cleanup
