+++ src/Uno.UWP/Graphics/Display/IDisplayInformationExtension.skia.cs
@@ -1,0 +1,17 @@
+﻿using System;
+namespace Windows.Graphics.Display;
+
+public interface IDisplayInformationExtension
+{
+	DisplayOrientations CurrentOrientation { get; }
+
+	uint ScreenHeightInRawPixels { get; }
+
+	uint ScreenWidthInRawPixels { get; }
+
+	float LogicalDpi { get; }
+
+	double RawPixelsPerViewPixel { get; }
+
+	ResolutionScale ResolutionScale { get; }
+}
